Set up submit endpoint
