+++ src/modules/site-v2/templates/tools/phenotype_database/phenotypeanalysisB.html
@@ -72,7 +72,16 @@
 {% block script %}
 {% from "_includes/select-trait.html"   import select_trait               %}
 {% from "_includes/strain-listing.html" import unpack_species_list        %}
+{% from "_scripts/submit-job.js"        import ajax_setup, def_submit_job %}
 <script>
+    {% include '_scripts/utils.js' %} {#/* defines: flash_message */#}
+    {{ def_submit_job('phenotype_database') }}
+
+
+    $(document).ready(function(){
+        // Set headers for AJAX requests
+        {{ ajax_setup(form.csrf_token._value()) }}
+    })
 
     /* Species */
     const species_list = {{ unpack_species_list(species_list, species_fields) }};
@@ -87,6 +96,21 @@
 
     function get_trait_id(id) {
         return document.getElementById(`${id}-selector`).dataset.trait
+    }
+
+    function submit() {
+        const trait_1 = get_trait_id('trait-1');
+
+        if (trait_1 == null) {
+            throw Error();
+        }
+
+        data = { 'label': '???', species, trait_1 }
+
+        submit_job(data)
+            .done(() => {
+                console.log('Submitted!');
+            });
     }
 </script>
 {% endblock %}
